requirements.txt
# Python dependencies
flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
python-dotenv==0.19.1
sqlalchemy==1.4.26
pyjwt==2.3.0
polyline==1.4.0